Reorganize benchmarks into folders

+++ benchmarks/Baselines.cpp
@@ -30,8 +30,8 @@ BENCHMARK(UINT64T_TO_STRING)
 static void DUMMY(benchmark::State& state) {
   for (auto _ : state) {
     size_t s;
-    for (size_t i = 0; i < state.range(0); ++i) {
-      for (size_t j = 0; j < state.range(0); ++j) {
+    for (int64_t i = 0; i < state.range(0); ++i) {
+      for (int64_t j = 0; j < state.range(0); ++j) {
         s += j;
       }
     }
